app/requirements.txt
fastapi==0.112.2
pydantic==2.8.2
requests==2.31.0
beautifulsoup4== 4.12.3
mangum==0.17.0
httpx==0.27.2
python-dotenv==1.0.1